Remove redundant features: App Sidebar, Marketplace Creator, and V1 browser
- Remove App Sidebar feature entirely (6 files deleted)
- Remove applyAppConfig methods from all managers
- Delete UserContentCreatorView and Create button from marketplace
- Rename MarketplaceBrowserViewV2 to MarketplaceBrowserView
- Delete deprecated V1 MarketplaceBrowserView
