+++ claude-maestro/MarketplaceManager.swift
@@ -1010,15 +1010,6 @@ class MarketplaceManager: ObservableObject {
         }
     }
 
-    // MARK: - App Configuration
-
-    /// Apply an app configuration to a session
-    /// This sets the session's plugin config to match the app's enabled plugins
-    func applyAppConfig(_ app: AppConfig, to sessionId: Int) {
-        let config = SessionPluginConfig(enabledPluginIds: app.enabledPluginIds)
-        sessionPluginConfigs[sessionId] = config
-        persistSessionConfigs()
-    }
 }
 
 // MARK: - Errors
